역할이 중복된 구조체 삭제
OutputView구조체랑 ResultView구조체의 역할이 중복되어
OutputView구조체를 삭제하고 ResultView에 포함시켰습니다.
