+++ LadderGame/LadderGame/ResultView.swift
@@ -21,4 +21,20 @@ struct ResultView {
         verticalLine = " "
     }
 
+    
+    
+    //플레이어이름 출력하는 함수
+    static func showPlayerName(playerName: Array<String>) {
+        var showName = String()
+        
+        for index in 0..<playerName.count {
+            var name = playerName[index]
+            while name.count < 6 {
+                name.insert(" ", at: name.endIndex)
+            }
+            showName = showName + name
+        }
+        print(showName)
+    }
+
 }
